Move handleProfileDownload from api handlers to handlers
