+++ web/srv/server.go
@@ -97,6 +97,7 @@ func NewServer(addr, templateDir, staticDir, uuid, controllerNamespace, webpackD
 	server.router.GET("/tap", handler.handleIndex)
 	server.router.GET("/top", handler.handleIndex)
 	server.router.GET("/routes", handler.handleIndex)
+	server.router.GET("/profiles/new", handler.handleProfileDownload)
 	server.router.ServeFiles(
 		"/dist/*filepath", // add catch-all parameter to match all files in dir
 		filesonly.FileSystem(server.staticDir))
@@ -111,7 +112,6 @@ func NewServer(addr, templateDir, staticDir, uuid, controllerNamespace, webpackD
 	server.router.GET("/api/services", handler.handleApiServices)
 	server.router.GET("/api/tap", handler.handleApiTap)
 	server.router.GET("/api/routes", handler.handleApiTopRoutes)
-	server.router.GET("/profiles/new", handler.handleProfileDownload)
 
 	return httpServer
 }
